moved _followPath to EntryShortcut to remove it from each of 3 child classes. Made UpdaterBase extend EntryShortcut including handled no context manager on initialize
